add team leader & individual team

+++ backend/backend/src/main/java/com/crimecat/backend/web/gametheme/service/MakerTeamService.java
@@ -1,17 +1,18 @@
 package com.crimecat.backend.web.gametheme.service;
 
-import com.crimecat.backend.bot.user.domain.DiscordUser;
+import com.crimecat.backend.auth.service.DiscordOAuth2UserService;
 import com.crimecat.backend.bot.user.domain.User;
 import com.crimecat.backend.bot.user.repository.UserRepository;
 import com.crimecat.backend.exception.ErrorStatus;
+import com.crimecat.backend.exception.ServiceException;
 import com.crimecat.backend.web.gametheme.domain.MakerTeam;
 import com.crimecat.backend.web.gametheme.domain.MakerTeamMember;
 import com.crimecat.backend.web.gametheme.dto.*;
 import com.crimecat.backend.web.gametheme.repository.MakerTeamMemberRepository;
 import com.crimecat.backend.web.gametheme.repository.MakerTeamRepository;
-import com.crimecat.backend.web.webUser.domain.WebUser;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -23,11 +24,24 @@ public class MakerTeamService {
     private final MakerTeamMemberRepository teamMemberRepository;
     private final UserRepository userRepository;
 
+    private final DiscordOAuth2UserService oAuth2UserService;
+
     public void create(String name) {
-        teamRepository.save(MakerTeam.builder()
+        create(name, oAuth2UserService.getLoginUserId(), false);
+    }
+
+    @Transactional
+    public UUID create(String name, UUID leader, boolean isIndividual) {
+        MakerTeam team = teamRepository.save(MakerTeam.builder()
                 .name(name)
-                .build()
-        );
+                .isIndividual(isIndividual)
+                .build());
+        teamMemberRepository.save(MakerTeamMember.builder()
+                .userId(leader)
+                .team(team)
+                .isLeader(true)
+                .build());
+        return team.getId();
     }
 
     public GetTeamsResponse get() {
@@ -46,16 +60,28 @@ public class MakerTeamService {
                 .build();
     }
 
+    public List<MakerTeamMember> getIndividualTeams(UUID leaderId) {
+        List<MakerTeamMember> teams = teamMemberRepository.findByUserIdAndIsLeader(leaderId, true);
+        return teams.stream().filter(v -> v.getTeam().isIndividual()).toList();
+    }
+
     @Transactional
     public void delete(UUID teamId) {
+        isTeamLeaderOrThrow(teamId);
         teamRepository.findById(teamId).orElseThrow(ErrorStatus.TEAM_NOT_FOUND::asServiceException);
-        // TODO: 지울 수 있는지 확인
         teamRepository.deleteById(teamId);
     }
 
     @Transactional
-    public void addMembers(UUID teamId, List<MemberRequestDto> members) {
+    public void addMembers(UUID teamId, Set<MemberRequestDto> members) {
+        isTeamLeaderOrThrow(teamId);
+        if (members.size() == 0) {
+            return;
+        }
         MakerTeam team = teamRepository.findById(teamId).orElseThrow(ErrorStatus.TEAM_NOT_FOUND::asServiceException);
+        if (team.isIndividual()) {
+            team.setIndividual(false);
+        }
         for (MemberRequestDto m : members) {
             MakerTeamMember.MakerTeamMemberBuilder builder = MakerTeamMember.builder().team(team);
             String name = m.getName();
@@ -67,13 +93,7 @@ public class MakerTeamService {
                 builder.userId(user.getId());
                 // 팀 멤버 이름 명시하지 않았고, 유저 존재할 때 유저 정보에서 이름 가져오기
                 if (name == null) {
-                    WebUser webUser = user.getWebUser();
-                    if (webUser != null) {
-                        name = webUser.getNickname();
-                    } else {
-                        DiscordUser discordUser = user.getDiscordUser();
-                        name = discordUser.getName();
-                    }
+                    name = user.getName();
                 }
                 // userId로 중복 유저 확인
                 if (team.getMembers().stream().anyMatch(v -> user.getId().equals(v.getUserId()))) {
@@ -88,36 +108,58 @@ public class MakerTeamService {
                 throw ErrorStatus.TEAM_MEMBER_ALREADY_REGISTERED.asServiceException();
             }
             builder.name(name);
-            MakerTeamMember newMember = builder.build();
             teamMemberRepository.save(builder.build());
         }
+        teamRepository.save(team);
     }
 
     @Transactional
     public void deleteMembers(UUID teamId, List<String> deletedMembers) {
-        MakerTeam team = teamRepository.findById(teamId).orElseThrow(ErrorStatus.TEAM_NOT_FOUND::asServiceException);
-        // 효율적으로 처리하기 위해 정렬
-        deletedMembers.sort(Comparator.naturalOrder());
-        List<MakerTeamMember> members = team.getMembers();
-        members.sort(Comparator.comparing(MakerTeamMember::getId));
-        Iterator<MakerTeamMember> iterator = members.iterator();
-        for (String deletedId : deletedMembers) {
-            if (!iterator.hasNext()) {
-                break;
+        if (deletedMembers.size() == 0) {
+            return;
+        }
+        try {
+            isTeamLeaderOrThrow(teamId);
+            MakerTeam team = teamRepository.findById(teamId).orElseThrow(ErrorStatus.TEAM_NOT_FOUND::asServiceException);
+            // 효율적으로 처리하기 위해 정렬
+            deletedMembers.sort(Comparator.naturalOrder());
+            List<MakerTeamMember> members = team.getMembers();
+            members.sort(Comparator.comparing(MakerTeamMember::getId));
+            Iterator<MakerTeamMember> iterator = members.iterator();
+            for (String deletedId : deletedMembers) {
+                if (!iterator.hasNext()) {
+                    break;
+                }
+                // 삭제하고자 하는 ID와 같은 ID 찾기
+                MakerTeamMember member = iterator.next();
+                String id = member.getId().toString();
+                while (id.compareTo(deletedId) < 0 && iterator.hasNext()) {
+                    member = iterator.next();
+                    id = member.getId().toString();
+                }
+                // 같으면 제거
+                if (deletedId.equals(id)) {
+                    teamMemberRepository.deleteById(member.getId());
+                    members.remove(member);
+                }
             }
-            // 삭제하고자 하는 ID와 같은 ID 찾기
-            MakerTeamMember member = iterator.next();
-            String id = member.getId().toString();
-            while (id.compareTo(deletedId) < 0 && iterator.hasNext()) {
-                member = iterator.next();
-                id = member.getId().toString();
-            }
-            // 같으면 제거
-            if (deletedId.equals(id)) {
-                teamMemberRepository.deleteById(member.getId());
-                members.remove(member);
+        } catch (ServiceException e) {
+            UUID userId = oAuth2UserService.getLoginUserId();
+            if (e.getStatus() == HttpStatus.FORBIDDEN && deletedMembers.contains(userId.toString())) {
+                teamMemberRepository.deleteById(userId);
+            } else {
+                throw e;
             }
         }
-        teamRepository.save(team);
+
+    }
+
+    private void isTeamLeaderOrThrow(UUID teamId) {
+        UUID userId = oAuth2UserService.getLoginUserId();
+        MakerTeamMember makerTeamMember = teamMemberRepository.findByUserIdAndTeamId(userId, teamId)
+                .orElseThrow(ErrorStatus.FORBIDDEN::asServiceException);
+        if (!makerTeamMember.isLeader()) {
+            throw ErrorStatus.FORBIDDEN.asServiceException();
+        }
     }
 }
